// src/types/definition.ts
export interface Book {
  no: number | null | undefined;
  title: string | null | undefined;
  author: string | null | undefined;
  rigDate: string | null | undefined;
  location: string | null | undefined;
  publisher: string | null | undefined;
  category: string | null | undefined;
  oneline: string | null | undefined;
  review: string | null | undefined;
  image: string | null | undefined;
  publishDate: string | null | undefined;
  writer: string | null | undefined;
}